+++ internal/_adapter/gateways/discordgo/channel.go
