+++ internal/_adapter/gateways/discordgo/discord.go
