+++ internal/_adapter/gateways/discordgo/role.go
