+++ internal/_adapter/gateways/discordgo/user.go
