+++ internal/_adapter/gateways/gs/read.go
