+++ internal/application/consts.go
