+++ internal/application/create/create_roles.go
@@ -6,6 +6,9 @@ import (
 
 	"github.com/murasame29/hackathon-util/cmd/config"
 	"github.com/murasame29/hackathon-util/internal/application"
+	"github.com/murasame29/hackathon-util/internal/datasource"
+	"github.com/murasame29/hackathon-util/internal/datasource/csv"
+	"github.com/murasame29/hackathon-util/internal/datasource/sheet"
 	"github.com/murasame29/hackathon-util/internal/discord"
 	"golang.org/x/sync/errgroup"
 )
@@ -68,13 +71,13 @@ func (o *CreateRolesOptions) Run() error {
 	}
 
 	cr := newCreateRole(discord)
-	var result *application.ReadDataSourceResult
+	var result *datasource.ReadDataSourceResult
 
 	switch o.dataSourceMode {
 	case application.DataSourceModeFile:
-		result, err = application.NewDataSourceCSV(o.FilePath).Read()
+		result, err = csv.NewDataSource(o.FilePath).Read()
 	case application.DataSourceModeGoogleSheet:
-		result, err = application.NewDataSourceGoogleSheets(o.SheetID, o.Range).Read()
+		result, err = sheet.NewDataSource(o.SheetID, o.Range).Read()
 	}
 
 	if err != nil {
